requirements.txt
fastapi
uvicorn
strands-agents
openai
python-dotenv
